pyproject.toml
[tool.poetry]
name = "nwms-scheduler"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ansi2html = "1.8.0"
anyio = "4.0.0"
apscheduler = "3.10.1"
asyncio = "3.4.3"
authlib = "1.2.1"
bcrypt = "4.0.1"
blinker = "1.6.2"
build = "1.0.3"
cachecontrol = "0.13.1"
cachelib = "0.9.0"
certifi = "2023.5.7"
cffi = "1.16.0"
charset-normalizer = "3.1.0"
cleo = "2.1.0"
click = "8.1.3"
colorama = "0.4.6"
crashtest = "0.4.1"
cryptography = "41.0.3"
dash = "2.13.0"
dash-core-components = "2.0.0"
dash-html-components = "2.0.0"
dash-table = "5.0.0"
distlib = "0.3.8"
dnspython = "2.3.0"
dulwich = "0.21.7"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
flask = "2.2.5"
flask-caching = "2.0.2"
flask-login = "0.6.2"
flask-wtf = "1.1.1"
h11 = "0.14.0"
httpcore = "0.17.3"
httpx = "0.24.1"
idna = "3.4"
importlib-metadata = "7.0.1"
installer = "0.7.0"
itsdangerous = "2.1.2"
jaraco-classes = "3.3.0"
jinja2 = "3.1.2"
keyring = "24.3.0"
markupsafe = "2.1.3"
mongoengine = "0.27.0"
more-itertools = "10.2.0"
msgpack = "1.0.7"
nest-asyncio = "1.5.6"
packaging = "23.1"
pexpect = "4.9.0"
pkginfo = "1.9.6"
platformdirs = "3.11.0"
plotly = "5.15.0"
poetry = "1.7.1"
poetry-core = "1.8.1"
poetry-plugin-export = "1.6.0"
ptyprocess = "0.7.0"
pycparser = "2.21"
pymongo = "4.4.0"
pyproject-hooks = "1.0.0"
python-dotenv = "1.0.0"
pytz = "2023.3"
pywin32-ctypes = "0.2.2"
rapidfuzz = "3.6.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
retrying = "1.3.4"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.0"
tenacity = "8.2.2"
tomlkit = "0.12.3"
trove-classifiers = "2024.1.8"
typing-extensions = "4.7.1"
tzdata = "2023.3"
tzlocal = "5.0.1"
urllib3 = "2.0.3"
virtualenv = "20.25.0"
werkzeug = "2.2.3"
wtforms = "3.0.1"
zipp = "3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
